Moves clear_pycondor_environment_variables to pycondor/utils.py
