--- pycondor/tests/utils.py
@@ -1,8 +1,0 @@
-
-import os
-
-
-def clear_pycondor_environment_variables():
-    # Unset any pycondor directory environment variables
-    for i in ['submit', 'output', 'error', 'log']:
-        os.environ['PYCONDOR_{}_DIR'.format(i.upper())] = ''
