+++ pycondor/utils.py
@@ -89,3 +89,9 @@ def string_rep(obj, quotes=False):
         obj_str = str(obj)
 
     return quote + obj_str + quote
+
+
+def clear_pycondor_environment_variables():
+    # Unset any pycondor directory environment variables
+    for i in ['submit', 'output', 'error', 'log']:
+        os.environ['PYCONDOR_{}_DIR'.format(i.upper())] = ''
